Move valid_exporter? to common lib.
