+++ sdk/lib/opentelemetry/sdk/trace/export/simple_span_processor.rb
@@ -29,7 +29,7 @@ module OpenTelemetry
           # @return [SimpleSpanProcessor]
           # @raise ArgumentError if the span_exporter is nil.
           def initialize(span_exporter)
-            raise ArgumentError, "exporter #{span_exporter.inspect} does not appear to be a valid exporter" unless Internal.valid_exporter?(span_exporter)
+            raise ArgumentError, "exporter #{span_exporter.inspect} does not appear to be a valid exporter" unless Common::Utilities.valid_exporter?(span_exporter)
 
             @span_exporter = span_exporter
           end
